Got rid of ifs in MessageHandlerImpl. Implemented a handler selection using the strategy pattern.
