+++ src/main/java/com/nelolik/savecountbot/handler/callback/CallbackData.java
@@ -1,0 +1,6 @@
+package com.nelolik.savecountbot.handler.callback;
+
+public interface CallbackData {
+    String CREATE_RECORD_BTN_DATA = "NEW_RECORD";
+    String ADD_COUNT_BTN_DATA = "ADD_COUNT";
+}
