+++ src/main/java/com/nelolik/savecountbot/handler/callback/CallbackHandler.java
@@ -1,0 +1,8 @@
+package com.nelolik.savecountbot.handler.callback;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
+
+public interface CallbackHandler {
+    SendMessage handle(CallbackQuery callbackQuery);
+}
