+++ src/main/java/com/nelolik/savecountbot/handler/callback/CallbackStringConstants.java
@@ -1,0 +1,7 @@
+package com.nelolik.savecountbot.handler.callback;
+
+public class CallbackStringConstants {
+    public static final String TEXT_ENTER_NEW_RECORD_NAME = "Enter new record name";
+    public static final String TEXT_ERROR_COMMAND_PROCESSING = "An error occurred during command processing. Try again.";
+    public static final String TEXT_ENTER_COUNT = "Enter count you want to save";
+}
