+++ src/main/java/com/nelolik/savecountbot/handler/context/ContextHandlerImpl.java
@@ -1,5 +1,6 @@
-package com.nelolik.savecountbot.handler;
+package com.nelolik.savecountbot.handler.context;
 
+import com.nelolik.savecountbot.handler.context.ContextHandler;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
