+++ src/main/java/com/nelolik/savecountbot/handler/context/ContextPhase.java
@@ -1,0 +1,8 @@
+package com.nelolik.savecountbot.handler.context;
+
+public enum ContextPhase {
+    SAVE_COUNT_REQUESTED,
+    RECORD_NAME_FOR_SAVE_COUNT_ENTERED,
+    NEW_RECORD_REQUESTED,
+    DELETE_RECORD_REQUESTED
+}
