+++ src/main/java/com/nelolik/savecountbot/handler/message/ApiCommands.java
@@ -1,0 +1,10 @@
+package com.nelolik.savecountbot.handler.message;
+
+public interface ApiCommands {
+    String COMMAND_HELLO = "/hello";
+    String COMMAND_NEW_RECORD = "/new_record";
+    String COMMAND_NEW_COUNT = "/new_count";
+    String COMMAND_LIST_OF_RECORDS = "/list_of_records";
+    String COMMAND_DELETE_RECORD = "/delete_record";
+    String NO_COMMAND = "no_command";
+}
