+++ src/main/java/com/nelolik/savecountbot/handler/message/TextHandler.java
@@ -1,0 +1,8 @@
+package com.nelolik.savecountbot.handler.message;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Message;
+
+public interface TextHandler {
+    SendMessage handle(Message message);
+}
